authenticate-user/requirements.txt
psycopg2-binary
pyotp
qrcode

